+++ tests/test_basic.py
@@ -1,6 +1,5 @@
 import asyncio
 import inspect
-import platform
 import sys
 from functools import _CacheInfo, partial
 from typing import Callable
@@ -152,8 +151,8 @@ async def test_alru_cache_dict_not_shared(check_lru: Callable[..., None]) -> Non
     assert ret1 == ret2
 
     assert (
-        coro1._LRUCacheWrapper__cache[1].fut.result()  # type: ignore[attr-defined]
-        == coro2._LRUCacheWrapper__cache[1].fut.result()  # type: ignore[attr-defined]
+        coro1._LRUCacheWrapper__cache[1].task.result()  # type: ignore[attr-defined]
+        == coro2._LRUCacheWrapper__cache[1].task.result()  # type: ignore[attr-defined]
     )
     assert coro1._LRUCacheWrapper__cache != coro2._LRUCacheWrapper__cache  # type: ignore[attr-defined]
     assert coro1._LRUCacheWrapper__cache.keys() == coro2._LRUCacheWrapper__cache.keys()  # type: ignore[attr-defined]
@@ -200,10 +199,6 @@ async def test_alru_cache_method() -> None:
     )
 
 
-@pytest.mark.xfail(
-    sys.version_info[:2] == (3, 9) and platform.python_implementation() != "PyPy",
-    reason="#511",
-)
 async def test_alru_cache_classmethod() -> None:
     class A:
         offset = 3